# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CSVImportConfirmRequestDTO:
    properties:
      expenses:
        items:
          $ref: '#/definitions/dto.CreateExpenseRequestDTO'
        type: array
      row_number:
        type: integer
    type: object
...
  dto.CategoryResponseDTO:
    properties:
      color:
        type: string
      created_at:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.CreateCategoryRequestDTO:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        type: string
    required:
    - color
    - name
    type: object
...
  dto.ExpenseResponseDTO:
    properties:
      added_by:
        type: string
      amount:
        type: number
      category:
        type: string
      comment:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: integer
      paid_by_card:
        description: 'Deprecated: kept for backward compatibility'
        type: boolean
      payment_method:
        type: string
      tags:
        items:
          $ref: '#/definitions/dto.TagResponseDTO'
        type: array
      type:
        type: string
      updated_at:
        type: string
      vendor:
        $ref: '#/definitions/dto.VendorResponseDTO'
    type: object
...
  dto.TagResponseDTO:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.UpdateCategoryRequestDTO:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Smart expense tracking application API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Expenso API
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category with the provided data
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a category
      tags:
      - categories
    get:
      consumes:
      - application/json
      description: Get a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a category by ID
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update an existing category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a category
      tags:
      - categories
  /expenses:
    get:
      consumes:
      - application/json
      description: Get a list of all expenses with optional date range filtering
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Create a new expense with the provided data
      parameters:
      - description: Expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ExpenseResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new expense
      tags:
      - expenses
  /expenses/{expense_id}/tags:
    get:
      description: Get all tags for a specific expense
      parameters:
      - description: Expense ID
        in: path
        name: expense_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TagResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get tags by expense
      tags:
      - tags
  /expenses/{expense_id}/tags/{tag_id}:
    delete:
      description: Remove a tag from an expense
      parameters:
      - description: Expense ID
        in: path
        name: expense_id
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: tag_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Remove tag from expense
      tags:
      - tags
    post:
      description: Add a tag to an expense
      parameters:
      - description: Expense ID
        in: path
        name: expense_id
        required: true
        type: integer
      - description: Tag ID
        in: path
        name: tag_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Add tag to expense
      tags:
      - tags
  /expenses/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an expense
      tags:
      - expenses
    get:
      consumes:
      - application/json
      description: Get a single expense by its ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExpenseResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an expense by ID
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Update an existing expense by ID
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated expense data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExpenseRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ExpenseResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an expense
      tags:
      - expenses
  /expenses/actual:
    get:
      consumes:
      - application/json
      description: Get expenses excluding salary entries for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get actual expenses (excluding salary)
      tags:
      - expenses
  /expenses/averages:
    get:
      consumes:
      - application/json
      description: Calculate average expenses per category and vendor type per month
        for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get average expenses per category and vendor per month
      tags:
      - expenses
  /expenses/balance:
    get:
      consumes:
      - application/json
      description: Get balance summary with total earnings, expenses, and balance
        for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get balance summary (earnings vs expenses)
      tags:
      - expenses
  /expenses/by-category:
    get:
      consumes:
      - application/json
      description: Get expenses filtered by category and optional date range, sorted
        by amount (highest first)
      parameters:
      - description: Category name to filter by
        in: query
        name: category
        required: true
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get expenses by category
      tags:
      - expenses
  /expenses/check-duplicates:
    get:
      consumes:
      - application/json
      description: Check for existing expenses with the same amount within a date
        range
      parameters:
      - description: Expense amount
        in: query
        name: amount
        required: true
        type: number
      - description: Expense date (YYYY-MM-DD)
        in: query
        name: date
        required: true
        type: string
      - description: 'Day range for duplicate check (default: 2)'
        in: query
        name: day_range
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check for duplicate expenses
      tags:
      - expenses
  /expenses/earnings:
    get:
      consumes:
      - application/json
      description: Get salary entries (earnings) for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get earnings (salary entries)
      tags:
      - expenses
  /expenses/export/csv:
    get:
      consumes:
      - application/json
      description: Export expenses filtered by card payment and "he" as CSV with vendor
        type columns
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV file
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Export expenses as CSV
      tags:
      - expenses
  /expenses/import/csv/confirm:
    post:
      consumes:
      - application/json
      description: Confirm and create expenses from a CSV row
      parameters:
      - description: Expenses to import
        in: body
        name: import_data
        required: true
        schema:
          $ref: '#/definitions/dto.CSVImportConfirmRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/dto.ExpenseResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Confirm and import CSV row
      tags:
      - expenses
  /expenses/import/csv/preview:
    post:
      consumes:
      - application/json
      description: Parse and preview CSV data for import validation
      parameters:
      - description: CSV data to preview
        in: body
        name: csv_data
        required: true
        schema:
          $ref: '#/definitions/dto.CSVImportRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CSVImportPreviewDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Preview CSV import
      tags:
      - expenses
  /incomes:
    get:
      consumes:
      - application/json
      description: Get a list of all incomes with optional date range filtering
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IncomeResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all incomes
      tags:
      - incomes
    post:
      consumes:
      - application/json
      description: Create a new income record
      parameters:
      - description: Income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/dto.CreateIncomeRequestDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.IncomeResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new income
      tags:
      - incomes
  /incomes/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an income record by ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an income
      tags:
      - incomes
    get:
      consumes:
      - application/json
      description: Get a specific income by its ID
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get income by ID
      tags:
      - incomes
    put:
      consumes:
      - application/json
      description: Update an existing income record
      parameters:
      - description: Income ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated income data
        in: body
        name: income
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateIncomeRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an income
      tags:
      - incomes
  /incomes/source/{source}:
    get:
      consumes:
      - application/json
      description: Get incomes filtered by source
      parameters:
      - description: Income source
        in: path
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.IncomeResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get incomes by source
      tags:
      - incomes
  /incomes/summary:
    get:
      consumes:
      - application/json
      description: Get total income amount and count for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.IncomeSummaryDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get income summary
      tags:
      - incomes
...
swagger: "2.0"